Cleanup locutus-node + streamline server
