+++ crates/core/src/bin/freenet.rs
@@ -18,8 +18,7 @@ async fn run_local(config: NodeConfig) -> Result<(), DynError> {
     let ip = config.address;
     let executor = Executor::from_config(config).await?;
     let socket: SocketAddr = (ip, port).into();
-    // freenet_core::local_node::run_local_node(executor, socket).await
-    todo!()
+    freenet_core::server::local_node::run_local_node(executor, socket).await
 }
 
 fn main() -> Result<(), DynError> {
