+++ crates/core/src/lib.rs
@@ -7,6 +7,8 @@ mod operations;
 mod resource_manager;
 mod ring;
 mod router;
+#[cfg(feature = "websocket")]
+pub mod server;
 pub mod util;
 
 pub type WrappedContract = locutus_runtime::prelude::WrappedContract;
@@ -16,8 +18,6 @@ type DynError = Box<dyn std::error::Error + Send + Sync + 'static>;
 
 // exports:
 pub use crate::config::Config;
-#[cfg(feature = "websocket")]
-pub use client_events::websocket::WebSocketProxy;
 pub use client_events::{
     combinator::ClientEventsCombinator, AuthToken, BoxedClient, ClientEventsProxy, ClientId,
     HostResult, OpenRequest,
