+++ crates/core/src/server.rs
@@ -1,19 +1,18 @@
 pub(crate) mod errors;
 mod http_gateway;
-pub(crate) mod web_handling;
+pub(crate) mod path_handlers;
 
-use freenet_core::{locutus_runtime::ContractKey, AuthToken, ClientId, HostResult};
 use freenet_stdlib::{
     client_api::{ClientError, ClientRequest, HostResponse},
     prelude::ContractInstanceId,
 };
-pub use http_gateway::HttpGateway;
+use locutus_runtime::ContractKey;
 
-type DynError = Box<dyn std::error::Error + Send + Sync + 'static>;
+use crate::{AuthToken, ClientId, HostResult};
 
 #[derive(Debug)]
 #[allow(clippy::large_enum_variant)]
-enum ClientConnection {
+pub(crate) enum ClientConnection {
     NewConnection {
         notifications: tokio::sync::mpsc::UnboundedSender<HostCallbackResult>,
         assigned_token: Option<(AuthToken, ContractInstanceId)>,
@@ -26,7 +25,7 @@ enum ClientConnection {
 }
 
 #[derive(Debug)]
-enum HostCallbackResult {
+pub(crate) enum HostCallbackResult {
     NewId {
         id: ClientId,
     },
@@ -44,10 +43,27 @@ enum HostCallbackResult {
 pub mod local_node {
     use std::net::{IpAddr, SocketAddr};
 
-    use freenet_core::{ClientEventsProxy, Executor, ExecutorError, OpenRequest, WebSocketProxy};
+    use axum::Router;
     use freenet_stdlib::client_api::{ClientRequest, ErrorKind};
+    use tokio::sync::mpsc;
+    use tower_http::trace::TraceLayer;
 
-    use crate::{DynError, HttpGateway};
+    use crate::{
+        client_events::websocket::WebSocketProxy, ClientEventsProxy, DynError, Executor,
+        ExecutorError, OpenRequest,
+    };
+
+    use super::http_gateway::HttpGateway;
+
+    const PARALLELISM: usize = 10; // TODO: get this from config, or whatever optimal way
+
+    async fn serve(socket: SocketAddr, router: Router) {
+        tracing::info!("listening on {}", socket);
+        axum::Server::bind(&socket)
+            .serve(router.into_make_service())
+            .await
+            .expect("Server unable to start");
+    }
 
     pub async fn run_local_node(
         mut executor: Executor,
@@ -62,11 +78,16 @@ pub mod local_node {
             }
             _ => {}
         }
-        let (mut http_handle, router) = HttpGateway::as_router(&socket);
-        let _ws_handle = WebSocketProxy::as_upgrade(socket, router).await?;
+        let (proxy_request_sender, request_to_server) = mpsc::channel(PARALLELISM);
+        let router = HttpGateway::as_router(&socket);
+        let (mut ws_proxy, router) =
+            WebSocketProxy::as_router(router, request_to_server, proxy_request_sender);
+
         // FIXME: use combinator
         // let mut all_clients =
         //    ClientEventsCombinator::new([Box::new(ws_handle), Box::new(http_handle)]);
+        serve(socket, router.layer(TraceLayer::new_for_http())).await;
+
         loop {
             let OpenRequest {
                 id,
@@ -74,7 +95,7 @@ pub mod local_node {
                 notification_channel,
                 token,
                 ..
-            } = http_handle.recv().await?;
+            } = ws_proxy.recv().await?;
             tracing::trace!(cli_id = %id, "got request -> {request}");
 
             let res = match *request {
@@ -84,21 +105,21 @@ pub mod local_node {
                         .await
                 }
                 ClientRequest::DelegateOp(op) => {
-                    let attested_contract = token.and_then(|token| {
-                        http_handle.attested_contracts.get(&token).map(|(t, _)| t)
-                    });
+                    let attested_contract = token
+                        .and_then(|token| ws_proxy.attested_contracts.get(&token).map(|(t, _)| t));
                     executor.delegate_request(op, attested_contract)
                 }
                 ClientRequest::Disconnect { cause } => {
                     if let Some(cause) = cause {
                         tracing::info!("disconnecting cause: {cause}");
                     }
-                    if let Some(rm_token) = http_handle
+                    // todo: token must live for a bit to allow reconnections
+                    if let Some(rm_token) = ws_proxy
                         .attested_contracts
                         .iter()
                         .find_map(|(k, (_, eid))| (eid == &id).then(|| k.clone()))
                     {
-                        http_handle.attested_contracts.remove(&rm_token);
+                        ws_proxy.attested_contracts.remove(&rm_token);
                     }
                     continue;
                 }
@@ -107,11 +128,11 @@ pub mod local_node {
 
             match res {
                 Ok(res) => {
-                    http_handle.send(id, Ok(res)).await?;
+                    ws_proxy.send(id, Ok(res)).await?;
                 }
                 Err(err) => {
                     tracing::error!("{err}");
-                    http_handle
+                    ws_proxy
                         .send(
                             id,
                             Err(ErrorKind::Unhandled {
