+++ crates/fdev/Cargo.toml
@@ -9,6 +9,7 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/freenet/locutus"
 
 [dependencies]
+anyhow = "1"
 bs58 = { workspace = true }
 # byteorder = "1"
 clap = { workspace = true, features = ["derive", "env"] }
