--- crates/locutus-node/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "locutus"
-version = "0.0.4"
-edition = "2021"
-rust-version = "1.71.1"
-publish = true
-description = "Locutus P2P network main node implementation"
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/freenet/locutus"
-
-[[bin]]
-name = "ldt"
-path = "src/bin/dev-tool.rs"
-
-[dependencies]
-anyhow = "1"
-async-trait = "0.1"
-axum = { version = "0.6", default-features = false, features = ["ws", "tower-log", "matched-path", "headers", "query"] }
-bincode = "1"
-bytes = "1"
-byteorder = "1"
-clap = { version = "4", features = ["derive", "env"] }
-cookie = "0.17"
-futures = "0.3.21"
-serde = "1"
-serde_json = "1"
-tar = "0.4.38"
-tokio = { version = "1", features = ["rt-multi-thread", "sync", "macros"] }
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter", "fmt"] }
-xz2 = "0.1"
-tower-http = { version = "0.4", features = ["trace", "fs"] }
-
-# internal:
-freenet-core = { path = "../core", version = "0.0.4" }
-fdev = { path = "../fdev", version = "0.0.4" }
-freenet-stdlib = { path = "../stdlib", version = "0.0.4" }
